ref(explore): Move explore toolbar styles
Move the styles into the explore/components folder to prepare shared components.
