+++ static/app/views/explore/toolbar/toolbarSortBy.tsx
@@ -7,6 +7,12 @@ import {Tooltip} from 'sentry/components/core/tooltip';
 import {t} from 'sentry/locale';
 import type {Sort} from 'sentry/utils/discover/fields';
 import {
+  ToolbarHeader,
+  ToolbarLabel,
+  ToolbarRow,
+  ToolbarSection,
+} from 'sentry/views/explore/components/toolbar/styles';
+import {
   useExploreFields,
   useExploreGroupBys,
   useExploreMode,
@@ -17,8 +23,6 @@ import {
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import {useSortByFields} from 'sentry/views/explore/hooks/useSortByFields';
 import {Tab, useTab} from 'sentry/views/explore/hooks/useTab';
-
-import {ToolbarHeader, ToolbarLabel, ToolbarRow, ToolbarSection} from './styles';
 
 export function ToolbarSortBy() {
   const mode = useExploreMode();
